# documentation/api/openapi.yaml
openapi: 3.0.0
info:
  title: My Daily Planner API
  description: |
    This is the My Daily Planner API documentation.
  version: 0.0.1
servers:
- url: https://it-project-2021-back-end.herokuapp.com/api
  description: Heroku production deployment
- url: http://localhost:4040/api
  description: Local development server
paths:
...
  /users/{userId}:
    get:
      tags:
      - users
      summary: Get user
      operationId: getUserId
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 6123778d576d702575e97590
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "5XX":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Users
    put:
      tags:
      - users
      summary: Update user
      operationId: putUserId
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 6123778d576d702575e97590
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "5XX":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Users
    delete:
      tags:
      - users
      summary: Delete user
      operationId: deleteUserId
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 6123778d576d702575e97590
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "5XX":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Users
...
components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: 6123778d576d702575e97590
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          example: johndoe@gmail.com
        password_hash:
          type: string
          example: $2a$10$BQtK9eybr0mbPC2R4cxeaeVx7WK.Z0G1TjI0wKqBGzfnIE6Wa.Dxu
        contacts:
          type: array
          example:
          - 6123785eec27a723da40feae
          - 6123827fd394f951583f0578
          items:
            type: string
        meetings:
          type: array
          example:
          - 613efb07bcfb116d50665acb
          - 613efb689b7c8e6ef8fdc4e6
          items:
            type: string
      example:
        _id: 6123778d576d702575e97590
        first_name: John
        last_name: Doe
        email: johndoe@gmail.com
        password_hash: $2a$10$BQtK9eybr0mbPC2R4cxeaeVx7WK.Z0G1TjI0wKqBGzfnIE6Wa.Dxu
        contacts:
        - 6123785eec27a723da40feae
        - 6123827fd394f951583f0578
        meetings:
        - 613efb07bcfb116d50665acb
        - 613efb689b7c8e6ef8fdc4e6
      xml:
        name: Users
...
    Error:
      required:
      - message
      - stack
      type: object
      properties:
        message:
          type: string
        stack:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...